+++ CHANGELOG.md
@@ -15,6 +15,8 @@ for further information about branching and tagging conventions.
 ### Added
 - Ability to save a few keys to key storage
 
+## [1.13.1] - 2020-07-16
+
 ## [1.13.1-rc.0] - 2020-06-02
 ### Changed
 - Subject size from 256 to 1000 
@@ -300,7 +302,8 @@ for further information about branching and tagging conventions.
 
 ## [1.3.0] - 2019-03-01
 
-[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.13.1-rc.0...HEAD
+[Unreleased]: https://github.com/tokend/new-js-sdk/compare/1.13.1...HEAD
+[1.13.1]: https://github.com/tokend/new-js-sdk/compare/1.13.1-rc.0...1.13.1
 [1.13.1-rc.0]: https://github.com/tokend/new-js-sdk/compare/1.13.0...1.13.1-rc.0
 [1.13.0]: https://github.com/tokend/new-js-sdk/compare/1.13.0-rc.0...1.13.0
 [1.13.0-rc.0]: https://github.com/tokend/new-js-sdk/compare/1.12.2...1.13.0-rc.0